BOCM-358 - refactor lpa-questionnaire-states and actions
