+++ apps/api/src/server/app/state-machine/household-appeal.machine.test.js
@@ -3,14 +3,14 @@ import test from 'ava';
 import sinon from 'sinon';
 import { transitionState } from './transition-state.js';
 import inspectorActionsService from './inspector.actions.js';
-import lpaQuestionnaireActions from './lpa-questionnaire.actions.js';
+import lpaQuestionnaireActionsService from './lpa-questionnaire-actions.service.js';
 
 const lpaQuestionnaireStub = sinon.stub();
 const inspectorSendBookingStub = sinon.stub();
 const notifyAppellantOfDecisionStub = sinon.stub();
 
 test.before('sets up mocking of actions', () => {
-	sinon.stub(lpaQuestionnaireActions, 'sendLpaQuestionnaire').callsFake(lpaQuestionnaireStub);
+	sinon.stub(lpaQuestionnaireActionsService, 'sendLpaQuestionnaire').callsFake(lpaQuestionnaireStub);
 	sinon.stub(inspectorActionsService, 'sendEmailToAppellantWithSiteVisitBooking').callsFake(inspectorSendBookingStub);
 	sinon.stub(inspectorActionsService, 'sendEmailToLPAAndAppellantWithDeciion').callsFake(notifyAppellantOfDecisionStub);
 });
